Move jose validator options to own file
