+++ validate/josev2/validator.go
@@ -22,9 +22,6 @@ type Validator struct {
 	allowedClockSkew   time.Duration                              // Optional.
 }
 
-// Option is how options for the Validator are set up.
-type Option func(*Validator)
-
 // CustomClaims defines any custom data / claims wanted.
 // The Validator will call the Validate function which
 // is where custom validation logic can be defined.
@@ -68,35 +65,6 @@ func New(
 	}
 
 	return v, nil
-}
-
-// WithAllowedClockSkew is an option which sets up the allowed
-// clock skew for the token. Note that in order to use this
-// the expected claims Time field MUST not be time.IsZero().
-// If this option is not used clock skew is not allowed.
-func WithAllowedClockSkew(skew time.Duration) Option {
-	return func(v *Validator) {
-		v.allowedClockSkew = skew
-	}
-}
-
-// WithCustomClaims sets up a function that returns the object
-// CustomClaims that will be unmarshalled into and on which
-// Validate is called on for custom validation. If this option
-// is not used the Validator will do nothing for custom claims.
-func WithCustomClaims(f func() CustomClaims) Option {
-	return func(v *Validator) {
-		v.customClaims = f
-	}
-}
-
-// WithExpectedClaims sets up a function that returns the object
-// used to validate claims. If this option is not used a default
-// jwt.Expected object is used which only validates token time.
-func WithExpectedClaims(f func() jwt.Expected) Option {
-	return func(v *Validator) {
-		v.expectedClaims = f
-	}
 }
 
 // ValidateToken validates the passed in JWT using the jose v2 package.
